The branch was cut from a stale local main, 148 commits behind, so this merge was
larger than it looked. Two files conflicted.

DirectGpuTensorEngine.Audio.cs — a genuine semantic conflict. main refactored the
GPU audio ops to return a deferred download (DeferTensorResult +
outBuf.RelinquishOwnership) instead of FinishGpuOp + new Tensor. My side added the
DifferentiableOps.RecordUnary calls those overrides were missing. Taking either
side alone would have been wrong: "mine" reverts main's deferred-download
optimization, "theirs" drops the gradient fix and leaves the GPU path silently
non-differentiable. Resolved by keeping main's deferred result AND recording it.
Recording the deferred tensor is safe — RecordUnary stores the reference and
materialization happens on access, which is when the backward reads it.

BackwardFunctions.cs — not a semantic conflict at all. main normalized the repo to
LF per .gitattributes (*.cs text eol=lf) while this file was previously UTF-8
BOM + CRLF, so every line differed and git conflicted all 7000 of them. main's
SpectrogramBackward was itself untouched by those 148 commits. Resolved by taking
main's normalized file wholesale and re-splicing my three additions onto it, which
keeps main's line-ending normalization and all its other work intact.

Verified the resolution rather than assuming it: the diff of the resolved
BackwardFunctions.cs against origin/main is +240/-7, and all 7 removed lines are
exactly the old ISTFT-delegating body being replaced. Nothing of main's was lost.

Also normalized TapeCompletenessTests.cs to LF, which had picked up CRLF.

Tests: 23/23 across the spectral, completeness, audio and mel suites on the merged
tree. Build clean.

+++ src/AiDotNet.Tensors/Engines/DirectGpu/Vulkan/VulkanAudioKernels.cs
@@ -41,7 +41,7 @@ void main() {
 ";
 
     public static string MuLawDecoding => Header + @"
-layout(set = 0, binding = 0) readonly buffer I { float input_[]; };
+layout(set = 0, binding = 0) readonly buffer I { int input_[]; };
 layout(set = 0, binding = 1) writeonly buffer O { float output_[]; };
 layout(push_constant) uniform P { int length; int quantizationChannels; };
 void main() {
@@ -49,7 +49,7 @@ void main() {
     if (gid >= length) return;
     if (quantizationChannels < 2) { output_[gid] = 0.0; return; }
     float mu = float(quantizationChannels - 1);
-    float q = input_[gid];
+    float q = float(input_[gid]);
     float y = (q / mu) * 2.0 - 1.0;
     float sgn = float(y > 0.0) - float(y < 0.0);
     output_[gid] = sgn * (pow(1.0 + mu, abs(y)) - 1.0) / mu;
